glados-pycromanager/glados_pycromanager/requirements.txt
bioio==3.0.0
bioio-ome-tiff
bioio-ome-zarr
bioimageio.core==0.8.0
csbdeep==0.7.3
google-auth-oauthlib==1.0.0
importlib-resources==5.12.0
markdown_captions
napari==0.5.6
ndtiff==2.1.0
numpy==2.0# Used to be 1.26.0
numba==0.61.2
opencv-python==4.11.0.86
pdoc
pycromanager==1.0.2
pyqtgraph==0.13.3
PyQtWebEngine==5.15.7
slack==0.0.2
slackeventsapi==3.0.3
tensorflow_intel==2.18.0
zarr==2.17.0